backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
asyncpg
alembic
pydantic>=2.0
pydantic-settings
python-jose[cryptography]
httpx
